Split of shared code to base
